+++ Shared/Components/RazorSEN0219.razor
@@ -8,9 +8,10 @@
         {
             return new ElementPin[]
             {
-                new ElementPin(name: "5V", x: 44, y: 181, signals: new PinSignalInfo[] { }, number: 1),
-                new ElementPin(name: "gnd", x: 37, y: 181, signals: new PinSignalInfo[] { }, number: 2),
-                new ElementPin(name: "analog_out", x: 60, y: 181, signals: new PinSignalInfo[] { }, number: 3),
+                new ElementPin(name: "gnd", x: 200.9, y: 68.67, signals: new PinSignalInfo[] { PinSignalInfo.GND() }, number: 1),
+                new ElementPin(name: "5V", x: 200.87, y: 79.615, signals: new PinSignalInfo[] { PinSignalInfo.VCC() }, number: 2),
+                new ElementPin(name: "analog_out", x: 200.91, y: 58.793, signals: new PinSignalInfo[] { }, number: 3),
+                new ElementPin(name: "NF", x: 200.91, y: 58.793, signals: new PinSignalInfo[] { }, number: 3)
             };
         }
     }
